Do not process meeting requests in sent items folder
grommunio-web and grommunio-sync process meeting requests and create
tentative appointments in the calendar. Processing a meeting request
in the sent items folder has an undesired side effect when creating
a meeting request in someone else's calendar. It will also create
a tentative appointment in the user's calendar even though the
user doesn't participate in the meeting.

References: grommunio-web#254